aspects: support finding schemas at end of path
Add a SchemaAt method that returns the schemas that may be at the end of
the path. Also add a Type method that returns a constant representation
of the type (Int, String, etc). These can be used to check there is any
value that may satisfy the schemas under two paths or if these will
always be incompatible.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>