Tests: detect if nested tests need to be executed based on change files
This change allows to detect changes in nested files even if there is
not a dependency with labeler workflow.

The dependency with the labeler workflow was creating duplicated runs,
so it has been removed and we need to detect manually the changes to
make sure the first time a nested test/lib is changed the nested tests
are triggered.