overlord: fix prerequisite base snap silently not being installed due to conflict (#13589)
* o/snapstate: add test for spurious snapd downgrade

The test sets up a scenario where a snap, upon refresh, needs to pull
in core22 base snap, but the change incorrectly conflicts with a parallel
refresh of snapd that is artifically crippled to not record its version
attribute.

Two new test-only snap IDs are added to fake store backend:

- some-snap-with-new-base-id this snap uses core22 as base
- snapd-without-version-id this snap creates a snapd without
  the version field. In fact the suffix -without-version-id
  causes this behavior for other snaps should we choose to add
  more.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* o/snapstate: add snap-setup to TestFinishRestartGeneratesSnapdWrappersOnCore

This is needed by the new non-test logic.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* o/snapstate: apply fix for snapd downgrade from Samuele

As seen on https://paste.ubuntu.com/p/QmDxmh2Mvg/

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* o/snapstate: explain snapd version special logic

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

---------

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>