o/i/a/common: allow groups in path patterns
Path patterns can be thought of as a base path concatenated with either
a suffix or a group over multiple suffixes.

The list of suffixes (by precedence) is:

1. ``           (no suffix has highest priority)
2. `/*.ext`     (particular file extension in base path directory)
3. `/**/*.ext`  (particular file extension in any subdirectory of base path)
4. `/.*`        (dotfiles in base path directory)
5. `/**/*.`     (dotfiles in any subdirectory of base path)
6. `/*`         (any file in base path directory)
7. `/**`        (any file in any subdirectory of base path)

When a path is received, it is matched against existing rules, each of
which has a path pattern. If there are any groups in a rule's path
pattern, that pattern has been expanded (at time of rule being added)
into a pattern without groups for each option in the group of the
original path pattern.

When the path incoming path is matched against existing rules, it may
match more than one path pattern. If there is only one pattern which is
of the form with the highest precedence (of the set of matching
patterns), then that rule has precedence. Otherwise, the longest pattern
has precedence, as all suffixes of a given precedence matching a given
rule must be identical, so the base path of the longest pattern must be
longest, and thus the most specific.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>