o/i/a/common: allow multi-part file extensions in path patterns
In order to support multi-part file extensions, such as `.tar.gz`, in
path patterns, we need to change the way precedence is computed for
patterns in the same suffix-precedence category.

For example, the following have the same suffix type (`/**/*.ext`), but
have equal length, so the old way of computing precedence by pattern
length fails:

* `/foo/bar/baz/**/*.gz`
* `/foo/bar/**/*.tar.gz`

In the above, we see that the first rule has a more specific base path,
while the second rule has a more specific file extension. It is a bit
arbitrary whether base path or file extension is more important, but for
consistency's sake, we introduce a new principle to path matching:

**Within a given suffix-precedence category, base path specificity is
more important than suffix specificity.**

Now, (1) has higher precedence than (2), despite that the latter has a
longer total pattern:

1. `/foo/bar/x/**/*.gz`
2. `/foo/bar/**/*.tar.gz

The precedence relationship still holds between suffix categories:

1. `/foo/bar/*.gz`
2. `/foo/bar/**/*.tar.gz`
3. `/foo/bar/*`
4. `/foo/bar/**`

As before, (1) has precedence over (2), because it matches files with a
given extension in the base path directory, rather than files with a
(more specific) file extension in any subdirectory, since any
subdirectory is a lower-precedence suffix category than in-directory.
The rules 1-4 are ordered in terms of precedence.

Note: the following should never be compared, since they cannot match
the same path, but if they *were* to be compared, the precedence order
would be as follows:

1. `/foo`
2. `/foo/bar/*.gz`
3. `/foo/*.tar.gz`
4. `/foo/bar/baz/**/*.gz`
5. `/foo/bar/**/*.tar.gz`
6. `/foo/bar/baz/fizz/*`
7. `/foo/bar/baz/*`
8. `/foo/bar/baz/fizz/buzz/**`
9. `/foo/bar/baz/fizz/**`

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>