interfaces: make steam-support implicit on core (#13189)
This is a commit we've had in the Core Desktop snapd tree for a while that had not been proposed upstream.

We would like to be able to run the Steam snap on Core Desktop. We've got everything else working, so providing an implicit steam-support slot on core (specifically Core Desktop) was the last piece of the puzzle. Everything else is similar enough not to be a problem.

* interfaces: make steam-support implicit on core

* interfaces: only make steam-support implicit on desktop variants of Ubuntu Core

* interfaces: test the StaticInfo for steam-support

---------

Co-authored-by: Ken VanDine <ken.vandine@canonical.com>