o/i/a/common: replaced usage of `strings.CutSuffix`
The `strings.CutSuffix` function requires go 1.20, which is not yet
required for snapd.

Furthermore, using the `CutSuffix` function in this case means calling
`HasSuffix` with `"/**/*"`, then constructing a string with the
`"/**/*"` removed, then constructing a new string with the `"/**"`
re-appended. Not very efficient.

Instead, simply check if the pattern has the `"/**/*"` suffix, and if
so, strip only the final `"/*"`.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>