o/i/a/common: allow arbitrary nested groups in path patterns
The `ExpandPathPattern` function has been rewritten to allow arbitrary
usage of groups in path patterns, without restrictions on the number of
groups which may occur nor the depth to which groups may be nested.

Now, the only restrictions placed on path patterns are that they must
begin with a `'/'` character and must not use unescaped `'['`, `']'`, or
`'?'` characters, as the current way precedence is computed relies on
being able to fully expand all path patterns into literals and
wildcards, and expanding `'[a-z]`' classes is not yet supported. If
classes were not expanded, they could be treated similarly to a single
`'?'` character with more restrictions, but computing precedence between
two classes is difficult, if not impossible.

As before, a `'\'` in front of any special character causes it to be
treated as a literal.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>