tests/lib/prepare.sh: reflash during shutdown
Originally, to reflash from Classic to Ubuntu Core, we rebooted with a
grub entry that would use specific `root=` and `init=` script to reflash
the disk.

We do not need a full reboot most of the time and we can just instead
prepare an initramfs to switch when shutting down. Systemd will switch
to `/run/initramfs` and run `/shutdown` in there.

However, on 16.04, systemd is unmounting mounts which makes this not
as easy. So instead there we still reboot to grub, but instead of using
`init=` and mount the root disk, we provide an initrd which can do the
flashing, simplifying the method we have been using.