gadget/install: consider real partition size when creating partitions
When building the list of partitions to create on auto-installation,
we were not properly calculating the end of a partition if that
partition existed on the disk and it had a size
range ([min-size,size]) defined in the gadget. Take this possibility
into account by properly matching on disk partitions to those in the
gadget and using the size on disk for existing partitions.