usersession: Use the app name from .desktop file in notifications
Currently, when a notification is shown to the user to inform
they that there is an update for an open snap, the snap name is
used, which can be misleading to the user.

This PR uses the application name from the .desktop file, taking
into account the current locale of the user.