packaging: fix debian builds broken by github.com/snapcore/bolt
The effort to allow Debian to build without bolt support, have been
insufficient, as we have discovered that parts of dh_golang do not support go
build tags.

As an alternative that is brutal but works, remove the single offending file
that causes 'go list -f ...' to enumerate github.com/snapcore/bolt due to
missing -tags nobolt argument that debhelper does not provide.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>