osutil: extend nfs-based home to support cifs
NFS and CIFS share the same shortcoming, that on kernels older than 6.8 (but I
could be wrong on the version number), accessing files there causes apparmor to
require network permissions even though the process performing the access is
entirely unaware of the fact that network file system is in use.

Snapd contained a work-around for this behavior, with specific detection logic
looking for NFS file system mounted or declared at a specific place. Extend the
logic to handle NFS and CIFS equally, to give better support to SAMBA or
Windows Server-based home directories.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>