o/devicestate: do not try to unmount early kernel mount if there isn't one
Do not attempt to unmount early kernel mount point if one wasn't created during
preseeding. Normally, we conditionally add a setup-kernel-snap, which sets up
the mount where we have more device context information, but for the teardown in
mark-preseeded, it should be enough to just check if the mount was indeed
created.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>