sandbox/apparmor: add WriteInternalABIFiles
We need to pass a path to the ABI file to apparmor_parser and thus we need to
write the ABI text somewhere. The directory /var/lib/snapd/apparmor seems like
a good candidate since the profiles are in sub-directory and we don't have to
worry about creating any sub-directories ourselves.

As a different alternative we could use memfd to always write an ephemeral file
and pass the /proc/PID/fd/FD path but this seems as a very technical solution
that we don't quite need.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>