boot/seal_test.go: add cases where
As we are reverting some old behavior because of an ordering bug we
need to reintroduce some cases in a refactored test. If the shim and
grub do not change of path, we need to accept the chain new shim ->
old grub. This is because we get confused and mix up old and new grub
somewhere else that has not been fixed yet.