tests/lib/core-config.sh: remove root user data for pc snap in restore_pc_snap
This function, when used in combination with clean_snapd_lib, tried to
to make it appear as if the "x1" revision of "pc" was never installed.
However, subsiquent tests would fail if they removed what appeared to be
their last revision of "pc", since that results in an attempt to remove
the assumedly empty `/root/snap/pc`, which still contained
`/root/snap/pc/x1`.