boot: do not sort list of hashes for boot assets
This is needed to re-merge #13402.

Hashes for boot assets are in a list that is already predictable: if
there is only one hash, then the asset is not modified; if there are 2
hashes, then the first one is the old asset, and the second is the new
asset.

There is no need to re-sort the hashes. And we lose the semantic if we
do. We want to keep that semantic so we can prune impossible boot chains.