gadget: cleanup files removed by gadgets
When a gadget is installed, we compare the previous gadget to detect
removed files and remove them. Preserved files that are removed from
content list are kept.

This is important to be able to remove old boot firmware since some,
like shim, change behavior depending of files that are present.

This does not support directories. If the gadget install directories
with content, the files in the directories will not be removed.