interfaces/desktop: allow DBus Notifications access to plasmashell
apparmor in Ubuntu 24.04 LTS introduces an AppArmor profile for KDEs plasmashell
called "plasmashell". As such, the DBus Notifications object is now implemented
by a process that has the apparmor label "plasmashell" instead of
"unconfined" (since previously it did not have an AppArmor profile).

Extend the DBus rules for /org/freedesktop/Notifications to include this new
peer label to allow snaps to send notifications as before.

Fixes part of LP: #2056696

Signed-off-by: Alex Murray <alex.murray@canonical.com>