many: merge release 2.62 (#13749)
* gadget: if storage traits is zero sized file, assume traits do not exist (#13719)

This is a safe fallback, as no-existent storage traits is a valid usecase.

Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>

* tests: fix recovery-system-reboot install test that was being interrupted by a system reboot (#13736)

* tests/nested/manual/recovery-system-reboot: fix test that was being interrupted by a system reboot

In the case of an auto-refresh, the system would reboot. This resulted
in some cryptic failures. Make sure to wait for an auto-refresh to
happen before removing the recovery system, and make sure to wait for
the system to reboot once the auto-refresh has finished.

* tests/nested/manual/recovery-system-reboot: add an extra wait to continue waiting after reboot

* sandbox/apparmor: detect but ignore apparmor 4 (#13740)

Due to issues with incorrect behavior to mediate:

    stat /dev/mqueue

For applications governed by the profile that allows it via

    mqueue,

We cannot yet use apparmor 4, even if one is supported on the host. This does
impact userns mediation but it is better to have the old mediation and not
break snaps, than to have some new mediation in some cases and some unexpected
mediation in other cases.

Once the mqueue, issue is identified and we have updated bundled apparmor to a
stable release of apparmor 4, this patch can be reverted.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* release: 2.62

* NEWS: restore WIP items

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

---------

Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>
Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>
Co-authored-by: Ondra Kubik <ondrej.kubik@canonical.com>
Co-authored-by: Andrew Phelps <136256549+andrewphelpsj@users.noreply.github.com>
Co-authored-by: Zygmunt Bazyli Krynicki <zygmunt.krynicki@canonical.com>
Co-authored-by: ernestl <ernest.lotter@canonical.com>