cmd/snap-confine: avoid TOCTOU with /var/lib/snapd/hostfs
Coverity scan raised a possible TOCTOU issue with how we check SC_HOSTFS_DIR.
Specifically we did stat() followed by mkdir(). The issue wasn't grave as
mkdir() was set to fail on any error (that includes EEXIST). However, to ensure
a peace of mind, reverse the order of operation, such that we mkdir() first and
then verify ownership if a directory was indeed created. This was tagged as
https://cwe.mitre.org/data/definitions/367.html

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>