o/snapshotstate/backend: delay finalizer call during StreamTo
The `NewSnapshotExport` function sets the finalizer to call `Close()` on
the newly-created `SnapshotExport` insteance. This behavior was
introduced in commit 03135412a526ea26a84bf805f4a3d7c296515feb.

Calling `Close()` closes all of the snapshot file FDs. However, the
`StreamTo()` function may be reading from one of the snapshot files
after the last explicit reference to the snapshot export instance, if
the caller does not hold and use a reference to the `SnapshotExport`
instance after the call to `StreamTo`. Thus, the finalizer might be
called while that read is still occurring or about to happen, thus
causing read errors because the FD is closed.

This caused the `TestExportTwice` test to fail intermittently, if the
garbage collector happened to run at just the right moment.

This commit fixes the issue at hand by adding a `runtime.KeepAlive` on
the snapshot export instance after all the reads of the snapshot files
are complete, thus preventing them from being closed early by the
finalizer calling `Close()`.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>