tests: move ubuntu core os.query logic to prepare.sh
As in snapd to prepare ubuntu core first it is booted a classic image.
The logic needed to manage that fact has been moved from os.query to
prepare.sh

This change is needed because os.query is a generic tool that is used in
other projects and shouln't have logic specific to snapd project.