interfaces/builtin/system_observe: allow listing systemd units and their properties (#13766)
* interfaces/builtin/system_observe: allow listing systemd units and their properties

Allow listing systemd units and querying their properties. This enables service
plugging system-observe to obtain more information about the state of the
system. Specifically, this enables Prometheus' systemd_exporter to work.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* tests/main/interfaces-system-observe: verify dbus permissions

Update the test to exercise dbus permissions to call system and get properties
of the manager and its units.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* NEWS: add a note about changes to system-observe interface

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

---------

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>