notices: Ensure that two notices don't have the same lastDateTime
When a client wants to receive the notices using the /v2/notices
interface, it usually will first receive all the old
notifications, store the date and time of the last notification,
and in the next calls use that date and time plus 1 nanosecond
to avoid geting the same notices over and over again.

Unfortunately, if, due to chance, two notifications have the
same date and time, this scheme will fail, because after
receiving the first one, the call will return it, and when the
client asks for the next, it will filter from the date/time plus
one microsecond, so the second notice won't be sent.

This patch fixes this by ensuring that no notice has the same
date/time value for lastReceived, by adding one or more
nanoseconds when required.